Fixed Android safe area

+++ library/include/borealis/core/application.hpp
@@ -54,8 +54,8 @@ enum class InputType
 
 struct SafeAreaInsets
 {
-    int top, bottom;
-    int left, right;
+    float top, bottom;
+    float left, right;
 };
 
 class DebugLayer;
@@ -168,7 +168,7 @@ class Application
      * return false if no actifity to pop.
      */
     static bool popActivity(
-        TransitionAnimation animation = TransitionAnimation::FADE, std::function<void(void)> cb = [] {}, bool free = true);
+        TransitionAnimation animation = TransitionAnimation::FADE, std::function<void(void)> cb = [] { }, bool free = true);
 
     /**
      * Gives the focus to the given view
